--- a/.github/workflows/rust.yml
+++ b/.github/workflows/rust.yml
@@ -14,8 +14,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
-      - name: Install Dependencies
-        run: sudo apt install libxcb-shape0-dev libxcb-xfixes0-dev libxkbcommon-dev
+      - uses: Swatinem/rust-cache@v1
       - name: Build
         run: cargo build --verbose
       - name: Run tests
@@ -23,5 +22,5 @@ jobs:
   cargo-deny:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v2
-    - uses: EmbarkStudios/cargo-deny-action@v1
+      - uses: actions/checkout@v2
+      - uses: EmbarkStudios/cargo-deny-action@v1