--- a/Cargo.toml
+++ b/Cargo.toml
@@ -4,5 +4,5 @@ resolver = "2"
 members = [
   "clovers",
   "clovers-cli",
-  "clovers-gpu/shaders",
+  "clovers-gpu",
 ]