--- a/clovers-cli/Cargo.toml
+++ b/clovers-cli/Cargo.toml
@@ -10,20 +10,23 @@ path = "src/main.rs"
 
 [dependencies]
 # Internal
-clovers = { path = "../clovers", features = ["serde-derive", "stl", "traces"]}
+clovers = { path = "../clovers", features = ["serde-derive", "stl", "traces"], default-features = false }
+
 # External
-rayon = "1.5.3"
-rand = { version = "0.8.5", features = ["small_rng"], default-features = false }
-serde = { version = "1.0.137", features = ["derive"], default-features = false }
-serde_json = "1.0.81"
-image = "0.24.2"
 chrono = "0.4.19"
-humantime = "2.1.0"
-indicatif = { version = "0.16.2", features = ["rayon"] }
-clap = { version = "3.1.18", features = ["derive"] }
-spirv-builder = { git = "https://github.com/EmbarkStudios/rust-gpu", rev = "116bf9c4d413df7413ae931ee31b2777c7abf64e" }
-bytemuck = { version = "1.9.1", features = ["derive"] }
+clap = { version = "3.1.18", features = ["std", "derive"], default-features = false  }
 futures = "0.3.21"
+humantime = "2.1.0"
+image = { version =  "0.24.2", features = ["png"], default-features = false }
+indicatif = { version = "0.16.2", features = ["rayon"], default-features = false }
+rand = { version = "0.8.5", features = ["small_rng"], default-features = false }
+rayon = "1.5.3"
+serde = { version = "1.0.137", features = ["derive"], default-features = false }
+serde_json = { version = "1.0", features = ["alloc"], default-features = false }
 tracing = "0.1.34"
 tracing-subscriber = { version = "0.3.11", features = ["time"] }
-wgpu = { version = "0.12.0", features = ["spirv"] }
+
+# GPU specific
+bytemuck = { version = "1.9.1", features = ["derive"], default-features = false }
+spirv-builder = { git = "https://github.com/EmbarkStudios/rust-gpu", rev = "116bf9c4d413df7413ae931ee31b2777c7abf64e" }
+wgpu = { version = "0.12.0", features = ["spirv"], default-features = false }