--- a/clovers-cli/Cargo.toml
+++ b/clovers-cli/Cargo.toml
@@ -13,7 +13,6 @@ path = "src/main.rs"
 clovers = { path = "../clovers", features = ["serde-derive", "stl", "traces"], default-features = false }
 
 # External
-chrono = "0.4.19"
 clap = { version = "3.1.18", features = ["std", "derive"], default-features = false  }
 futures = "0.3.21"
 humantime = "2.1.0"
@@ -23,6 +22,7 @@ rand = { version = "0.8.5", features = ["small_rng", "getrandom"], default-featu
 rayon = "1.5.3"
 serde = { version = "1.0.137", features = ["derive"], default-features = false }
 serde_json = { version = "1.0", features = ["alloc"], default-features = false }
+time = { version = "0.3.9", default-features = false }
 tracing = "0.1.34"
 tracing-subscriber = { version = "0.3.11", features = ["time"] }