--- a/clovers-gpu/shaders/rust-toolchain.toml
+++ b/clovers-gpu/shaders/rust-toolchain.toml
@@ -1,4 +1,0 @@
-# Required for rust-gpu
-[toolchain]
-channel = "nightly-2022-04-11"
-components = ["rust-src", "rustc-dev", "llvm-tools-preview"]