--- a/clovers/Cargo.toml
+++ b/clovers/Cargo.toml
@@ -18,7 +18,7 @@ traces = ["tracing"]
 
 [dependencies]
 nalgebra = { version = "0.31.0", features = ["libm"], default-features = false }
-rand = { version = "0.8.5", features = ["small_rng"] }
+rand = { version = "0.8.5", features = ["small_rng"], default-features = false }
 serde = { version = "1.0.137", features = ["derive"], default-features = false, optional = true }
 stl_io = { version = "0.6.0", optional = true }
 tracing = { version = "0.1.34", optional = true }