--- a/clovers/src/colorize.rs
+++ b/clovers/src/colorize.rs
@@ -91,7 +91,8 @@ pub fn colorize(ray: &Ray, scene: &Scene, depth: u32, max_depth: u32, rng: &mut 
             let scattered = scattered.non_negative();
 
             // Blend it all together
-            emitted + scattered
+            // TODO: temporarily taking the average, as it fixes the overblown colors. This is (probably) not correct though!
+            (emitted + scattered) / 2.0
         }
     }
 }