--- a/clovers/src/random.rs
+++ b/clovers/src/random.rs
@@ -10,19 +10,21 @@ pub fn random_in_unit_sphere(rng: &mut SmallRng) -> Vec3 {
     // TODO: figure out a non-loop method
     // See https://github.com/RayTracing/raytracing.github.io/issues/765
     loop {
-        position = 2.0 * Vec3::new(rng.gen(), rng.gen(), rng.gen()) - Vec3::new(1.0, 1.0, 1.0);
+        position = Vec3::new(
+            rng.gen_range(-1.0..1.0),
+            rng.gen_range(-1.0..1.0),
+            rng.gen_range(-1.0..1.0),
+        );
         if position.magnitude_squared() >= 1.0 {
-            return position;
+            continue;
         }
+        return position;
     }
 }
 
 /// Internal helper. Use this for the more correct "True Lambertian" reflection
 pub fn random_unit_vector(rng: &mut SmallRng) -> Vec3 {
-    let a: Float = rng.gen_range(0.0..2.0 * PI);
-    let z: Float = rng.gen_range(-1.0..1.0);
-    let r: Float = (1.0 - z * z).sqrt();
-    Vec3::new(r * a.cos(), r * a.sin(), z)
+    random_in_unit_sphere(rng).normalize()
 }
 
 /// Internal helper.
@@ -31,18 +33,22 @@ pub fn random_in_unit_disk(rng: &mut SmallRng) -> Vec3 {
     // TODO: figure out a non-loop method
     // See https://github.com/RayTracing/raytracing.github.io/issues/765
     loop {
-        // TODO: understand this defocus disk thingy
-        position = 2.0 * Vec3::new(rng.gen(), rng.gen(), 0.0) - Vec3::new(1.0, 1.0, 0.0);
-        if position.dot(&position) >= 1.0 {
-            return position;
+        position = Vec3::new(
+            rng.gen_range(-1.0..1.0),
+            rng.gen_range(-1.0..1.0),
+            0.0, // z component zero
+        );
+        if position.magnitude_squared() >= 1.0 {
+            continue;
         }
+        return position;
     }
 }
 
 /// Internal helper.
 pub fn random_cosine_direction(rng: &mut SmallRng) -> Vec3 {
-    let r1 = rng.gen::<Float>();
-    let r2 = rng.gen::<Float>();
+    let r1: Float = rng.gen();
+    let r2: Float = rng.gen();
     let z = (1.0 - r2).sqrt();
 
     let phi = 2.0 * PI * r1;
@@ -53,14 +59,12 @@ pub fn random_cosine_direction(rng: &mut SmallRng) -> Vec3 {
 }
 
 /// Internal helper.
-pub fn random_to_sphere(radius: Float, distance_squared: Float, rng: &mut SmallRng) -> Vec3 {
-    let r1 = rng.gen::<Float>();
-    let r2 = rng.gen::<Float>();
-    let z = 1.0 + r2 * ((1.0 - radius * radius / distance_squared).sqrt() - 1.0);
-
-    let phi = 2.0 * PI * r1;
-    let x = phi.cos() * (1.0 - z * z).sqrt();
-    let y = phi.sin() * (1.0 - z * z).sqrt();
-
-    Vec3::new(x, y, z)
+pub fn random_in_hemisphere(normal: Vec3, rng: &mut SmallRng) -> Vec3 {
+    let in_unit_sphere: Vec3 = random_in_unit_sphere(rng);
+    if in_unit_sphere.dot(&normal) > 0.0 {
+        // In the same hemisphere as the normal
+        in_unit_sphere
+    } else {
+        -in_unit_sphere
+    }
 }